refactor: move time mode to lib
