+++ upace/lib/timeMode.ts
@@ -1,0 +1,5 @@
+
+export type TimeMode = {
+    inputModes: string[];
+    displayModes: string[];
+};
