+++ upace/screens/timeCalc/timeCalcUtils.ts
@@ -1,24 +1,20 @@
 import moment from "moment";
 import React from "react";
+import { TimeMode } from "../../lib/timeMode";
 
-export const getNewTimeSplit = () => ({ split: moment.duration(), ref: React.createRef() } as TimeSplit);
+export const getNewTimeSplit = () =>
+  ({ split: moment.duration(), ref: React.createRef() } as TimeSplit);
 export const initialValues = {
-    mode: { inputModes: ["mm", "ss", "ds"], displayModes: ["mm", "ss", "ds"] },
-    splits: [getNewTimeSplit()],
-} as TimeCalcValue;
+  mode: { inputModes: ["mm", "ss", "ds"], displayModes: ["mm", "ss", "ds"] },
+  splits: [getNewTimeSplit()],
+} as TimeCalcState;
 
-export type TimeCalcValue = {
-    mode : TimeMode,
-    splits : TimeSplit[]
-}
-
-export type TimeMode = {
-    inputModes: string[],
-    displayModes: string[],
-}
+export type TimeCalcState = {
+  mode: TimeMode;
+  splits: TimeSplit[];
+};
 
 export type TimeSplit = {
-    split: moment.Duration;
-    ref: React.RefObject<any>
-}
-
+  split: moment.Duration;
+  ref: React.RefObject<any>;
+};
